Refactor cart service

+++ backend/services/cart-service/cart/src/main/java/vn/uit/edu/msshop/cart/adapter/in/web/CartEventListener.java
@@ -1,5 +1,6 @@
 package vn.uit.edu.msshop.cart.adapter.in.web;
 
+import java.time.Instant;
 import java.util.List;
 
 import org.springframework.kafka.annotation.KafkaHandler;
@@ -8,6 +9,8 @@ import org.springframework.stereotype.Component;
 
 import lombok.RequiredArgsConstructor;
 import vn.uit.edu.msshop.cart.adapter.in.web.mapper.CartWebMapper;
+import vn.uit.edu.msshop.cart.adapter.out.event.documents.EventDocument;
+import vn.uit.edu.msshop.cart.adapter.out.event.repositories.EventDocumentRepository;
 import vn.uit.edu.msshop.cart.application.dto.command.DeleteCartItemCommand;
 import vn.uit.edu.msshop.cart.application.port.in.DeleteCartItemUseCase;
 import vn.uit.edu.msshop.cart.domain.event.OrderCreatedSuccess;
@@ -18,10 +21,13 @@ import vn.uit.edu.msshop.cart.domain.event.OrderCreatedSuccess;
 public class CartEventListener {
     private final DeleteCartItemUseCase deleteItemUseCase;
     private final CartWebMapper mapper;
+    private final EventDocumentRepository eventDocumentRepository;
     
     @KafkaHandler
     public void onOrderCreated(OrderCreatedSuccess orderCreatedSuccess) {
+        if(eventDocumentRepository.existsById(orderCreatedSuccess.eventId())) return;
         List<DeleteCartItemCommand> commands = mapper.toCommand(orderCreatedSuccess);
         deleteItemUseCase.deleteManyItems(commands);
+        eventDocumentRepository.save(EventDocument.builder().eventId(orderCreatedSuccess.eventId()).receiveAt(Instant.now()).build());
     }
 }
